[FIX] Moved check inactive thread function to the main root folder as Jay Asked
